Modified simulator to use chunks. Also fixed bug in circuit benchmark reader that had duplicate uninitialzed nodes (fixed such that all uninitialized nodes are removed)
Added block_width() accessor function to GPU_Data (needed to figure out offsets if two GPU_Datas have different sizes and both are being used at the same time).

Modified debug output to dump to file for simulator. Also removed pass.

Added unary function to check if a node is of unknown type.
Added stubs of a load/save function to dump a binary representation of Circuit to a file.

Main:
Switched vec to a  pointer with new/delete so that